graphql: split user code into files
