+++ src/typeorm/entities/users/graphql/index.ts
@@ -1,0 +1,4 @@
+export { login } from './user-login.mutation';
+export { register } from './user-register.mutation';
+export { UserLoginMutationTypeDefs } from './user-login.types';
+export { UserRegisterMutationTypeDefs } from './user-register.types';
